statistics moved to metrics package
