+++ platform-sdk/swirlds-common/src/main/java/com/swirlds/common/metrics/platform/AbstractDistributionMetric.java
@@ -18,7 +18,7 @@ package com.swirlds.common.metrics.platform;
 
 import com.swirlds.common.metrics.MetricConfig;
 import com.swirlds.common.metrics.platform.Snapshot.SnapshotEntry;
-import com.swirlds.common.statistics.StatsBuffered;
+import com.swirlds.common.metrics.statistics.StatsBuffered;
 import com.swirlds.common.utility.CommonUtils;
 import java.util.List;
