+++ platform-sdk/swirlds-common/src/main/java/com/swirlds/common/metrics/platform/DefaultStatEntry.java
@@ -27,7 +27,7 @@ import com.swirlds.common.metrics.MetricConfig;
 import com.swirlds.common.metrics.StatEntry;
 import com.swirlds.common.metrics.config.MetricsConfig;
 import com.swirlds.common.metrics.platform.Snapshot.SnapshotEntry;
-import com.swirlds.common.statistics.StatsBuffered;
+import com.swirlds.common.metrics.statistics.StatsBuffered;
 import java.util.List;
 import java.util.function.Consumer;
 import java.util.function.Supplier;
