+++ platform-sdk/swirlds-common/src/main/java/com/swirlds/common/metrics/statistics/internal/StatsBuffer.java
@@ -1,20 +1,21 @@
 /*
- * Copyright (C) 2017-2023 Hedera Hashgraph, LLC
+ * Copyright (C) 2023 Hedera Hashgraph, LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *      http://www.apache.org/licenses/LICENSE-2.0
+ *        http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
+ *
  */
 
-package com.swirlds.common.statistics.internal;
+package com.swirlds.common.metrics.statistics.internal;
 
 import static com.swirlds.common.utility.Units.NANOSECONDS_TO_SECONDS;
