+++ platform-sdk/swirlds-common/src/test/java/com/swirlds/common/metrics/LongAccumulatorTest.java
@@ -28,7 +28,7 @@ import static org.mockito.Mockito.spy;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 
-import com.swirlds.common.statistics.StatsBuffered;
+import com.swirlds.common.metrics.statistics.StatsBuffered;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
