+++ platform-sdk/swirlds-common/src/test/java/com/swirlds/common/metrics/StatEntryConfigTest.java
@@ -20,7 +20,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.Mockito.mock;
 
-import com.swirlds.common.statistics.StatsBuffered;
+import com.swirlds.common.metrics.statistics.StatsBuffered;
 import java.util.function.Consumer;
 import java.util.function.Function;
 import java.util.function.Supplier;
