+++ platform-sdk/swirlds-common/src/test/java/com/swirlds/common/metrics/StatEntryTest.java
@@ -23,8 +23,8 @@ import static com.swirlds.common.metrics.Metric.ValueType.VALUE;
 import static com.swirlds.common.metrics.MetricType.STAT_ENTRY;
 import static org.assertj.core.api.Assertions.assertThat;
 
-import com.swirlds.common.statistics.StatsBuffered;
-import com.swirlds.common.statistics.internal.StatsBuffer;
+import com.swirlds.common.metrics.statistics.StatsBuffered;
+import com.swirlds.common.metrics.statistics.internal.StatsBuffer;
 import java.util.function.Consumer;
 import java.util.function.Supplier;
 import org.junit.jupiter.api.DisplayName;
