+++ platform-sdk/swirlds-common/src/test/java/com/swirlds/common/metrics/platform/DefaultFunctionGaugeTest.java
@@ -29,7 +29,7 @@ import com.swirlds.common.metrics.FunctionGauge;
 import com.swirlds.common.metrics.IntegerGauge;
 import com.swirlds.common.metrics.Metric;
 import com.swirlds.common.metrics.platform.Snapshot.SnapshotEntry;
-import com.swirlds.common.statistics.StatsBuffered;
+import com.swirlds.common.metrics.statistics.StatsBuffered;
 import java.util.List;
 import java.util.function.Supplier;
 import org.junit.jupiter.api.Test;
